requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# AI and ML libraries
torch>=1.9.0
transformers==4.56.1
huggingface-hub==0.34.4

# PDF processing
PyMuPDF==1.24.5
PyPDF2==3.0.1
fpdf2==2.7.6

# Document processing
python-docx==1.1.0
markdown==3.5.1

# Web scraping and content extraction
newspaper3k==0.2.8
beautifulsoup4==4.13.4
requests==2.31.0

# YouTube processing
youtube-transcript-api==0.6.2

# Utilities
python-dateutil==2.9.0
tqdm==4.67.1